+++ rebook-members/src/main/java/com/be/rebook/members/controller/MembersController.java
@@ -53,15 +53,6 @@ public class MembersController {
         String majorsToUpdate = majorsDTO.getMajors();
         return ResponseEntity.ok().body(new BaseResponse<>(memberService.updateUserMajors(username, majorsToUpdate)));
     }
-
-    // 로그인 이후 마이페이지에서 바로 비밀번호 변경이라 이메일 인증 필요없음
-    // todo : 패스워드 수정 로직 테스트하기
-    @PatchMapping("/password")
-    public ResponseEntity<BaseResponse<Members>> updateUserPassword(@Auth MemberLoginInfo memberLoginInfo, @Valid UpdatePasswordDTO passwordDTO){
-        String username = memberLoginInfo.getUsername();
-        String passwordToUpdate = passwordDTO.getPassword();
-        return ResponseEntity.ok().body(new BaseResponse<>(memberService.updateUserPassword(username, passwordToUpdate)));
-    }
     //todo : 프로필 사진 변경 로직 테스트하기
     @PatchMapping("/profilePicture")
     public ResponseEntity<BaseResponse<Members>> updateUserPicture(@Auth MemberLoginInfo memberLoginInfo, MultipartFile picture){
